Removes seeding of database every time the test suite is ran
By seeding the database every time the test suite is ran, in addition to the setup that happens in each test, there ends up being *lots* of unneeded, unexpected, random data.

This extraneous, undependable data causes random test failures, in particular with the User model.

For example, in a given test, I would expect there to be a User.count of 2, and it ends up being 164.

Also, because of the Faker gem’s randomness, we would end up randomly getting zip codes and cities that “sometimes” overlap with the zip codes and cities being tested.  Again causing random failing tests.

I am not understanding why we need to seed the database before every test suite run, if the tests are set up correctly.  Tests themselves are supposed to set the stage for the test in question, IMO.