New GitHubUser model and table
GitHub associates their data with a GitHub user.  This user has a unique login and id.

For us to be able to associate statistics per user, we need to gather and leverage the user’s information.

This table will also allow us to link the GitHubUser to a `User` in out database.  This is not required for statistics to be calculated and presented, but is a nice touch, for users that provide us this information.