Pulls valid email regex into a constant, and increases strictness
This regex will now be used in multiple places, so I am pulling it into a reusable constant.

The original regex was really only enforcing the presence of a @ symbol.  As such, it was allowing an email address like this, that has nothing before the @ symbol, and no “.com, .org, etc.”:

“@gmail“

This new, more strict regex, enforces that:
- something come the @
- the presence of an @
- the presence of a “.something”, i.e. “.com”

This regex comes from the `validates_format_of` rails docs:

http://api.rubyonrails.org/classes/ActiveModel/Validations/HelperMethods.html#method-i-validates_format_of