ActiveJob perform_later does not accept a Struct wrapped in an Array
All of the tests were giving false positives, because in testing, `perform_now` is called in the background.

Also, during manual testing, perform_now was being used.

The catch is, perform_later does not accept an array wrapped Struct as an argument, despite the fact that is allows arrays.

As such, this logic was migrated from the controller, to the job itself.

Now the job is receiving a string as as an arg, which perform_later does accept.