Sets access to required forest env vars
By default, Forest sets the env vars to pull from the secrets.yml file (through the initializer).

Forest requires two sets of secrets:  one for development, and one for production.

The production secrets are account for through our kubernetes secrets process.

The development secrets need to be handled differently, as they are different than the values for prod.

As of this moment, I am choosing to protect those values, and not commit them to GitHub.

That could change, once I have tested this process with another developer, on their workstation.