Favors a role method over a belongs_to association
Reason being, with the belongs_to association, every single user record would *have to* have an associated role.

That is not the use case for us.  Standard users are authenticated, and do not need a named role.  Authentication grants them the access level they need.

The role will come into play for users that need increased access.  For example an admin, or board member, or super admin, etc., etc.