zip is valid if properly formatted or geocoded
If a zip is a number, that is either of the below formats:

‘12345’ or ‘12345-1234’

It is valid.

If it is not in one of those formats, and geocoding IS working, and geocoding cannot find the zip code, it is invalid.

If geocoding is not working, it accepts the passed zip code.