wrap some factory attrs in blocks to make them lazy
Laziness in a factory lets an attr be random. Without putting these in
Ruby blocks, FactoryGirl will call Faker for an attribute once when a
factory is first used and then use that value for every create() and
build() thereafter. This is helpful in development for being able to
generate a bunch of sample data in the database that looks different. It
will also ensure that the records for the test suite are randomish to
make sure tests aren't passing because of some underlying, but
unintentional consistency.

Signed-off-by: Robb Kidd <robb@thekidds.org>