remove user welcome from record callback to controller create action
This fixes a race condition between the web process creating the user
record and a worker process picking up the job to use the user record to
add them to SendGrid.

The race shows itself in an error from Sidekiq:

    ActiveJob::DeserializationError: Error while trying to deserialize
    arguments: Couldn't find User with 'id'=1234

The `welcome_user()` method is the thing scheduling a job that is trying
to find a user. Making the ActiveRecord lifecycle callbacks responsible
for business processes can be painful to test and maintain. This change
removes the welcoming of a user from the User model's callbacks and
makes the controller action that creates users responsible for welcoming
them.Asserting that the job is enqueued is then possible in the User
model test without the double-enqueue that would occur before: once when
the test user instance was created (saved and committed) and again when
calling the `welcome_user()` method to test it.

Summary of changes:

* User's after_create callback to welcome_user is removed.
* User's tests are updated to only confirm that the job is enqueued when
  welcome_user() is called.
* UsersController now calls welcome_user after a user is successfully
  saved to the database which avoids the race condition between save and
  job execution.
* UsersController tests are updated to:
  * assert that User's welcome_user is called during create controller
    action
  * share a common set of valid user_params
  * dry the tests slightly by using the valid_user_params with a removal
    to test invalid user params

Signed-off-by: Robb Kidd <robb@thekidds.org>