Added housing_provided column to locations table, defaulting to false
Created a new migration to add a housing_provided column to the locations table, with a default value of false. Also added housing_provided to the locations serializer so that it is rendered in the JSON returned from the API.

This commit is meant to solve issue #397. It will allow housing data to be stored in the back end for each location.